Eliminate build errors by Clang with debug mode
Modified the clang build to start with -Wall and -Wextra build options, so they do not overwrite the -Wno-unused-parameter flag.
This way, false positive unused parameter build errors can be avoided.

Added two new flags: -Wno-overloadedvirtual and -Wno-unused-private-field.

Also fixed two warnings.

+++ src/interpreter/ByteCode.h
@@ -428,7 +428,7 @@ class LoadThisBinding : public ByteCode {
 public:
     LoadThisBinding(const ByteCodeLOC& loc, const size_t dstIndex)
         : ByteCode(Opcode::LoadThisBindingOpcode, loc)
-        , m_dstIndex(dstIndex == REGULAR_REGISTER_LIMIT ? SIZE_MAX : dstIndex)
+        , m_dstIndex(dstIndex == REGULAR_REGISTER_LIMIT ? (uint16_t)SIZE_MAX : dstIndex)
     {
     }
